+++ CHANGELOG.md
@@ -50,6 +50,7 @@ Click up arrow to go back to TOC.
 - Add temp file save function. [issue #22](https://github.com/BingLingGroup/autosub/issues/22)
 - Add only audio fragments output. [issue #44](https://github.com/BingLingGroup/autosub/issues/44)
 - Add subtitles(first line) language detection powered by googletrans. [issue #40](https://github.com/BingLingGroup/autosub/issues/40)
+- Add http, https proxy support.(Set environment variables) [issue #17](https://github.com/BingLingGroup/autosub/issues/17)
 
 #### Changed(Unreleased)